CUDA Kmeans
Optimized CUDA code, better speedup.

+++ ParallelCUDA/KmeansCUDA.cu
@@ -11,46 +11,43 @@ static void CheckCudaErrorAux(const char* file, unsigned line, const char* state
 }
 #define CUDA_CHECK_ERROR(value) CheckCudaErrorAux(__FILE__,__LINE__, #value, value)
 
-__host__ void calculateMaxSSE(Point*points, Point* selectedCentroids, int numPoints, int numClusters, int gridSize, int blockSize, float& maxSSE) {
+__host__ void calculateMaxSSE(Point*points, Point* selectedCentroids, int numPoints, int numClusters, int gridSize, int blockSize, double& maxSSE) {
     Point* d_points;
     Point* d_currentCentroids;
-    float* d_maxSSE;
+    double* d_maxSSE;
     CUDA_CHECK_ERROR(cudaMalloc((void**)&d_points, numPoints * sizeof(Point)));
     CUDA_CHECK_ERROR(cudaMalloc((void**)&d_currentCentroids, numClusters * sizeof(Point)));
     CUDA_CHECK_ERROR(cudaMalloc((void**)&d_maxSSE, sizeof(double)));
     CUDA_CHECK_ERROR(cudaMemcpy(d_points, points, numPoints * sizeof(Point), cudaMemcpyHostToDevice));
     CUDA_CHECK_ERROR(cudaMemcpy(d_currentCentroids, selectedCentroids, numClusters * sizeof(Point), cudaMemcpyHostToDevice));
-
     CUDAcalculateMaxSSE<<<gridSize, blockSize>>>(d_points, d_currentCentroids, d_maxSSE, numPoints, numClusters);
-
-    CUDA_CHECK_ERROR(cudaMemcpy(&maxSSE, d_maxSSE, sizeof(float), cudaMemcpyDeviceToHost));
-
+    CUDA_CHECK_ERROR(cudaMemcpy(&maxSSE, d_maxSSE, sizeof(double), cudaMemcpyDeviceToHost));
     CUDA_CHECK_ERROR(cudaFree(d_points));
     CUDA_CHECK_ERROR(cudaFree(d_currentCentroids));
     CUDA_CHECK_ERROR(cudaFree(d_maxSSE));
 }
 
-__global__ void CUDAcalculateMaxSSE(Point* d_points, Point* d_currentCentroids, float* d_maxSSE, int numPoints, int numClusters){
+__global__ void CUDAcalculateMaxSSE(Point* d_points, Point* d_currentCentroids, double* d_maxSSE, int numPoints, int numClusters){
     int tid = blockIdx.x * blockDim.x + threadIdx.x;
     int pointsPerThread = (numPoints + blockDim.x - 1) / blockDim.x;
     int startPoint = tid * pointsPerThread;
     int endPoint = min(startPoint + pointsPerThread, numPoints);
 
-    float sse = 0.0f;
+    double sse = 0.0f;
     for (int pointIndex = startPoint; pointIndex < endPoint; pointIndex++) {
         Point dataPoint = d_points[pointIndex];
-        float minDistance = 0;
+        double minDistance = 0;
         minDistance += pow(dataPoint.x - d_currentCentroids[0].x,2);
         minDistance += pow(dataPoint.y - d_currentCentroids[0].y,2);
         minDistance += pow(dataPoint.z - d_currentCentroids[0].z,2);
         minDistance = sqrt(minDistance);
         for (int clusterId = 1; clusterId < numClusters; clusterId++) {
-            float distanceToCentroid = 0;
+            double distanceToCentroid = 0;
             distanceToCentroid += pow(dataPoint.x - d_currentCentroids[clusterId].x,2);
             distanceToCentroid += pow(dataPoint.y - d_currentCentroids[clusterId].y,2);
             distanceToCentroid += pow(dataPoint.z - d_currentCentroids[clusterId].z,2);
             distanceToCentroid = sqrt(distanceToCentroid);
-            minDistance = fminf(minDistance, distanceToCentroid);
+            minDistance = fmin(minDistance, distanceToCentroid);
         }
 
         // Accumulate the SSE for this data point
